followAlong/flappy_ai/pyproject.toml
[project]
name = "flappy-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flappy-bird-gymnasium>=0.4.0",
    "pyyaml>=6.0.2",
    "tensorflow>=2.20.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
